"add registry interface" (#6449)
* "add registry interface"

* "move function to registry"

* "rename with meaningful name"

* "add exposed layers"

* "fixed based on comments"

* "remove unsed comments"
